Added some clarification on ISO8601 dates
As suggested in https://github.com/fullcalendar/fullcalendar-site-static/pull/109, it might be a good idea to clarify the handling of basic and extended date and date-times. Users that have to write their own parser to comply with what Fullcalendar understands in contrast to what RFC5545 defines might find this information rather usefull.